requirements.txt
tqdm==4.64.1
networkx==3.0
shroomdk==1.0.2
python-dotenv==0.21.0
